See the extended description
Alright, here is what I am doing:
The actual map is separate from the SGArena file. The display name really belongs on a map, and as well as spawn locations, which is why I removed them from the initialize method - because that is for the SGWorld, the actual map, not the lobby.